rate products

+++ server/src/main/java/com/springboot/bookstore/repository/OrderDetailsRepository.java
@@ -5,5 +5,5 @@ import com.springboot.bookstore.entity.OrderDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
 }
